Add files via upload
This Blackjack game is a classic card game implemented using HTML, CSS, and JavaScript, providing an engaging and interactive experience with a visually appealing design. The game features a responsive interface with a green casino-themed background, allowing players to hit, stand, and reset the game. With dynamic score calculations and real-time updates, the game offers a seamless and enjoyable gameplay experience. The intuitive design ensures accessibility and ease of use, making it a standout project in my portfolio, showcasing my skills in front-end development and user interface design.